--- multimodal/src/autogluon/multimodal/models/fusion.py
@@ -1,580 +1,0 @@
-import logging
-from typing import List, Optional
-
-import torch
-import torch.nn.functional as F
-from torch import nn
-
-from ..constants import (
-    AUTOMM,
-    FEATURES,
-    LABEL,
-    LOGITS,
-    NER_ANNOTATION,
-    NER_TEXT,
-    TOKEN_WORD_MAPPING,
-    WEIGHT,
-    WORD_OFFSETS,
-)
-from .ft_transformer import CLSToken, FT_Transformer
-from .mlp import MLP
-from .utils import init_weights, run_model
-
-logger = logging.getLogger(AUTOMM)
-
-
-class BaseMultimodalFusionModel(nn.Module):
-    """
-    A base class to fuse different models' features (single-modal and multimodal).
-    """
-
-    def __init__(self):
-        super().__init__()
-
-    @property
-    def label_key(self):
-        return f"{self.prefix}_{LABEL}"
-
-    def forward(self, *args, **kwargs):
-        pass
-
-    def get_layer_ids(
-        self,
-    ):
-        """
-        Assign an id to each layer. Layer ids will be used in layer-wise lr decay.
-        Basically, id gradually increases when going from the output end to
-        the input end.
-
-        It assumes that each individual model has the "name_to_id" attribute storing
-        the already computed model's layer ids. This function only collects those layer ids.
-        It also add prefixes for each model's parameter names since the fusion model wraps
-        those individual models, making the name scope changed. Configuring the optimizer
-        requires a full name of each parameter.
-
-        The layers defined in this class, e.g., head, adapter,
-        and, fusion_mlp, have id 0.
-
-        Returns
-        -------
-        A dictionary mapping the layer names (keys) to their ids (values).
-        """
-        model_prefix = "model"
-        names = [n for n, _ in self.named_parameters()]
-
-        outer_layer_names = [n for n in names if not n.startswith(model_prefix)]
-        name_to_id = {}
-        logger.debug(f"outer layers are treated as head: {outer_layer_names}")
-        for n in outer_layer_names:
-            name_to_id[n] = 0
-
-        for i, per_model in enumerate(self.model):
-            per_model_prefix = f"{model_prefix}.{i}"
-            if not hasattr(per_model, "name_to_id"):
-                raise ValueError(f"name_to_id attribute is missing in model: {per_model.__class__.__name__}")
-            for n, layer_id in per_model.name_to_id.items():
-                full_n = f"{per_model_prefix}.{n}"
-                name_to_id[full_n] = layer_id
-
-        # double check each parameter has been assigned an id
-        for n in names:
-            assert n in name_to_id
-
-        return name_to_id
-
-
-class MultimodalFusionMLP(BaseMultimodalFusionModel):
-    """
-    Use MLP to fuse different models' features (single-modal and multimodal).
-    Specifically, it adapts the features of each model to specified dimensions,
-    concatenates the adapted features, and fuses the features through MLP.
-    """
-
-    def __init__(
-        self,
-        prefix: str,
-        models: list,
-        hidden_features: List[int],
-        num_classes: int,
-        adapt_in_features: Optional[str] = None,
-        activation: Optional[str] = "gelu",
-        dropout_prob: Optional[float] = 0.5,
-        normalization: Optional[str] = "layer_norm",
-        loss_weight: Optional[float] = None,
-    ):
-        """
-        Parameters
-        ----------
-        prefix
-            The fusion model's prefix
-        models
-            The individual models whose output features will be fused.
-        hidden_features
-            A list of integers representing the hidden feature dimensions. For example,
-            [512, 128, 64] indicates three hidden MLP layers with their corresponding output
-            feature dimensions.
-        num_classes
-            The number of classes.
-        adapt_in_features
-            Choice of how to adapt the features of each model. We now support
-            - min
-                Adapt all features to the minimum dimension. For example, if three models have
-                feature dimensions [512, 768, 64], it will linearly map all the features to
-                dimension 64.
-            - max
-                Adapt all features to the maximum dimension. For example, if three models have
-                feature dimensions are [512, 768, 64], it will linearly map all the features to
-                dimension 768.
-        activation
-            Name of activation function.
-        dropout_prob
-            Dropout probability.
-        normalization
-            Name of normalization function.
-        loss_weight
-            The weight of individual models. For example, if we fuse the features of ViT, CLIP, and BERT,
-            The loss will be computed as "loss = fusion_loss + loss_weight(vit_loss + clip_loss + bert_loss)".
-            Basically, it supports adding an auxiliary loss for each individual model.
-        """
-        super().__init__()
-        logger.debug("initializing MultimodalFusionMLP")
-        if loss_weight is not None:
-            assert loss_weight > 0
-        self.loss_weight = loss_weight
-        self.model = nn.ModuleList(models)
-        self.num_classes = num_classes
-
-        raw_in_features = [per_model.out_features for per_model in models]
-        if adapt_in_features is not None:
-            if adapt_in_features == "min":
-                base_in_feat = min(raw_in_features)
-            elif adapt_in_features == "max":
-                base_in_feat = max(raw_in_features)
-            else:
-                raise ValueError(f"unknown adapt_in_features: {adapt_in_features}")
-
-            self.adapter = nn.ModuleList([nn.Linear(in_feat, base_in_feat) for in_feat in raw_in_features])
-
-            in_features = base_in_feat * len(raw_in_features)
-        else:
-            self.adapter = nn.ModuleList([nn.Identity() for _ in range(len(raw_in_features))])
-            in_features = sum(raw_in_features)
-
-        assert len(self.adapter) == len(self.model)
-
-        fusion_mlp = []
-        for per_hidden_features in hidden_features:
-            fusion_mlp.append(
-                MLP(
-                    in_features=in_features,
-                    hidden_features=per_hidden_features,
-                    out_features=per_hidden_features,
-                    num_layers=1,
-                    activation=activation,
-                    dropout_prob=dropout_prob,
-                    normalization=normalization,
-                )
-            )
-            in_features = per_hidden_features
-        self.fusion_mlp = nn.Sequential(*fusion_mlp)
-        # in_features has become the latest hidden size
-        self.head = nn.Linear(in_features, num_classes)
-        # init weights
-        self.adapter.apply(init_weights)
-        self.fusion_mlp.apply(init_weights)
-        self.head.apply(init_weights)
-
-        self.out_features = in_features
-
-        self.prefix = prefix
-
-        self.name_to_id = self.get_layer_ids()
-        self.head_layer_names = [n for n, layer_id in self.name_to_id.items() if layer_id == 0]
-
-    def forward(
-        self,
-        batch: dict,
-    ):
-        """
-
-        Parameters
-        ----------
-        batch
-            A dictionary containing the input mini-batch data. The fusion model doesn't need to
-            directly access the mini-batch data since it aims to fuse the individual models'
-            output features.
-
-        Returns
-        -------
-        If "loss_weight" is None, it returns dictionary containing the fusion model's logits and
-        features. Otherwise, it returns a list of dictionaries collecting all the models' output,
-        including the fusion model's.
-        """
-        multimodal_features = []
-        output = {}
-        for per_model, per_adapter in zip(self.model, self.adapter):
-            per_output = run_model(per_model, batch)
-            multimodal_features.append(per_adapter(per_output[per_model.prefix][FEATURES]))
-            if self.loss_weight is not None:
-                per_output[per_model.prefix].update({WEIGHT: self.loss_weight})
-                output.update(per_output)
-
-        features = self.fusion_mlp(torch.cat(multimodal_features, dim=1))
-        logits = self.head(features)
-        fusion_output = {
-            self.prefix: {
-                LOGITS: logits,
-                FEATURES: features,
-            }
-        }
-        if self.loss_weight is not None:
-            fusion_output[self.prefix].update({WEIGHT: 1})
-            output.update(fusion_output)
-            return output
-        else:
-            return fusion_output
-
-
-class MultimodalFusionNER(BaseMultimodalFusionModel):
-    """
-    Use MLP to fuse different models' features (single-modal and multimodal) for NER.
-    Specifically, it adapts the features of each model to specified dimensions,
-    concatenates the adapted features, and fuses the features through MLP.
-    """
-
-    def __init__(
-        self,
-        prefix: str,
-        models: list,
-        hidden_features: List[int],
-        num_classes: int,
-        adapt_in_features: str = "max",
-        activation: Optional[str] = "gelu",
-        dropout_prob: Optional[float] = 0.5,
-        normalization: Optional[str] = "layer_norm",
-        loss_weight: Optional[float] = None,
-    ):
-        """
-        Parameters
-        ----------
-        prefix
-            The fusion model's prefix
-        models
-            The individual models whose output features will be fused.
-        hidden_features
-            A list of integers representing the hidden feature dimensions. For example,
-            [512, 128, 64] indicates three hidden MLP layers with their corresponding output
-            feature dimensions.
-        num_classes
-            The number of classes.
-        adapt_in_features
-            Choice of how to adapt the features of each model. We now support
-            - min
-                Adapt all features to the minimum dimension. For example, if three models have
-                feature dimensions [512, 768, 64], it will linearly map all the features to
-                dimension 64.
-            - max
-                Adapt all features to the maximum dimension. For example, if three models have
-                feature dimensions are [512, 768, 64], it will linearly map all the features to
-                dimension 768.
-        activation
-            Name of activation function.
-        dropout_prob
-            Dropout probability.
-        normalization
-            Name of normalization function.
-        loss_weight
-            The weight of individual models.
-        """
-        super().__init__()
-        logger.debug("initializing MultimodalFusionNER")
-
-        self.ner_model = None
-        self.tokenizer = None
-        other_models = []
-        for per_model in models:
-            if per_model.prefix != NER_TEXT:
-                other_models.append(per_model)
-            else:
-                self.ner_model = per_model
-                self.tokenizer = per_model.tokenizer
-        self.other_models = nn.ModuleList(other_models)
-        raw_in_features = [per_model.out_features for per_model in models if per_model.prefix != NER_TEXT]
-        if adapt_in_features is not None:
-            if adapt_in_features == "min":
-                base_in_feat = min(raw_in_features)
-            elif adapt_in_features == "max":
-                base_in_feat = max(raw_in_features)
-            else:
-                raise ValueError(f"unknown adapt_in_features: {adapt_in_features}")
-            self.adapter = nn.ModuleList([nn.Linear(in_feat, base_in_feat) for in_feat in raw_in_features])
-            in_features = base_in_feat * len(raw_in_features)
-        else:
-            self.adapter = nn.ModuleList([nn.Identity() for _ in range(len(raw_in_features))])
-            in_features = sum(raw_in_features)
-        assert len(self.adapter) == len(self.other_models)
-        fusion_mlp = []
-        for per_hidden_features in hidden_features:
-            fusion_mlp.append(
-                MLP(
-                    in_features=in_features,
-                    hidden_features=per_hidden_features,
-                    out_features=per_hidden_features,
-                    num_layers=1,
-                    activation=activation,
-                    dropout_prob=dropout_prob,
-                    normalization=normalization,
-                )
-            )
-            in_features = per_hidden_features
-        self.fusion_mlp = nn.Sequential(*fusion_mlp)
-        self.head = nn.Linear(in_features + self.ner_model.out_features, num_classes)
-
-    @property
-    def label_key(self):
-        return f"{NER_TEXT}_{LABEL}"
-
-    def forward(
-        self,
-        batch: dict,
-    ):
-        """
-        Parameters
-        ----------
-        batch
-            A dictionary containing the input mini-batch data. The fusion model doesn't need to
-            directly access the mini-batch data since it aims to fuse the individual models'
-            output features.
-        Returns
-        -------
-        It returns dictionary containing the fusion model's logits and features.
-        """
-        multimodal_features = []
-        output = {}
-        ner_output = run_model(self.ner_model, batch)
-        for per_model, per_adapter in zip(self.other_models, self.adapter):
-            per_output = run_model(per_model, batch)
-            multimodal_features.append(per_adapter(per_output[per_model.prefix][FEATURES]))
-
-        features = self.fusion_mlp(torch.cat(multimodal_features, dim=1))
-        features = features.unsqueeze(dim=1).repeat(1, ner_output[self.ner_model.prefix][FEATURES].size()[1], 1)
-        features = torch.cat((ner_output[self.ner_model.prefix][FEATURES], features), dim=-1)
-
-        logits = self.head(features)
-        logits_label = torch.argmax(F.log_softmax(logits, dim=-1), dim=-1)
-        fusion_output = {
-            self.prefix: {
-                LOGITS: logits,
-                FEATURES: features,
-                NER_ANNOTATION: logits_label,
-                TOKEN_WORD_MAPPING: ner_output[self.ner_model.prefix][TOKEN_WORD_MAPPING],
-                WORD_OFFSETS: ner_output[self.ner_model.prefix][WORD_OFFSETS],
-            }
-        }
-
-        return fusion_output
-
-
-class MultimodalFusionTransformer(BaseMultimodalFusionModel):
-    """
-    Use Transformer to fuse different models' features (single-modal and multimodal).
-    Specifically, it adapts the features of each model to specified dimensions,
-    concatenates the adapted features, and fuses the features through Transformer.
-    """
-
-    def __init__(
-        self,
-        prefix: str,
-        models: list,
-        hidden_features: int,
-        num_classes: int,
-        n_blocks: Optional[int] = 0,
-        attention_n_heads: Optional[int] = 8,
-        attention_initialization: Optional[str] = "kaiming",
-        attention_normalization: Optional[str] = "layer_norm",
-        attention_dropout: Optional[str] = 0.2,
-        residual_dropout: Optional[str] = 0.0,
-        ffn_activation: Optional[str] = "reglu",
-        ffn_normalization: Optional[str] = "layer_norm",
-        ffn_d_hidden: Optional[str] = 192,
-        ffn_dropout: Optional[str] = 0.0,
-        prenormalization: Optional[bool] = True,
-        first_prenormalization: Optional[bool] = False,
-        kv_compression_ratio: Optional[float] = None,
-        kv_compression_sharing: Optional[str] = None,
-        head_activation: Optional[str] = "relu",
-        head_normalization: Optional[str] = "layer_norm",
-        adapt_in_features: Optional[str] = None,
-        loss_weight: Optional[float] = None,
-        additive_attention: Optional[bool] = False,
-        share_qv_weights: Optional[bool] = False,
-    ):
-        """
-        Parameters
-        ----------
-        prefix
-            The fusion model's prefix
-        models
-            The individual models whose output features will be fused.
-        hidden_features
-            A list of integers representing the hidden feature dimensions. For example,
-            [512, 128, 64] indicates three hidden MLP layers with their corresponding output
-            feature dimensions.
-        num_classes
-            The number of classes.
-        n_blocks
-            Number of the `FT_Transformer` blocks, which should be non-negative.
-        attention_n_heads
-            Number of attention heads in each `FT_Transformer` block, which should be positive.
-        attention_dropout
-            Dropout ratio for the Multi Headed Attention module.
-        attention_initialization
-            Weights initialization scheme for Multi Headed Attention module.
-        attention_normalization
-            Normalization policy for attention layers. "layer_norm" is a good default.
-        residual_dropout
-            Dropout ratio for the linear layers in FT_Transformer block.
-        ffn_d_hidden
-            Number of the hidden nodes of the linear layers in the Feed-Forward Network module.
-        ffn_dropout
-            Dropout ratio of the hidden nodes of the linear layers in the Feed-Forward Network module.
-        ffn_activation
-            Activation function type for the Feed-Forward Network module.
-        ffn_normalization
-            Normalization scheme of the Feed-Forward Network module.
-        prenormalization, first_prenormalization
-            Prenormalization to stabilize the training.
-        kv_compression_ratio
-            The compression ration to reduce the input sequence length.
-        kv_compression_sharing
-            If `true` the projections will share weights.
-        head_activation
-            Activation function type of the MLP layer.
-        head_normalization
-            Normalization scheme of the MLP layer.
-        adapt_in_features
-            Choice of how to adapt the features of each model. We now support
-            - min
-                Adapt all features to the minimum dimension. For example, if three models have
-                feature dimensions [512, 768, 64], it will linearly map all the features to
-                dimension 64.
-            - max
-                Adapt all features to the maximum dimension. For example, if three models have
-                feature dimensions are [512, 768, 64], it will linearly map all the features to
-                dimension 768.
-        loss_weight
-            The weight of individual models. For example, if we fuse the features of ViT, CLIP, and BERT,
-            The loss will be computed as "loss = fusion_loss + loss_weight(vit_loss + clip_loss + bert_loss)".
-            Basically, it supports adding an auxiliary loss for each individual model.
-        additive_attention
-            If 'true' the transformer will use additive attention with linear complexity to sequence length.
-        share_qv_weights
-            if 'true', then value and query transformation parameters are shared in additive attention.
-        """
-        super().__init__()
-        logger.debug("initializing MultimodalFusionTransformer")
-        if loss_weight is not None:
-            assert loss_weight > 0
-
-        self.loss_weight = loss_weight
-        self.model = nn.ModuleList(models)
-
-        raw_in_features = [per_model.out_features for per_model in models]
-
-        if adapt_in_features == "min":
-            base_in_feat = min(raw_in_features)
-        elif adapt_in_features == "max":
-            base_in_feat = max(raw_in_features)
-        else:
-            raise ValueError(f"unknown adapt_in_features: {adapt_in_features}")
-
-        self.adapter = nn.ModuleList([nn.Linear(in_feat, base_in_feat) for in_feat in raw_in_features])
-
-        in_features = base_in_feat
-
-        assert len(self.adapter) == len(self.model)
-
-        self.fusion_transformer = FT_Transformer(
-            d_token=in_features,
-            n_blocks=n_blocks,
-            attention_n_heads=attention_n_heads,
-            attention_dropout=attention_dropout,
-            attention_initialization=attention_initialization,
-            attention_normalization=attention_normalization,
-            ffn_d_hidden=ffn_d_hidden,
-            ffn_dropout=ffn_dropout,
-            ffn_activation=ffn_activation,
-            ffn_normalization=ffn_normalization,
-            residual_dropout=residual_dropout,
-            prenormalization=prenormalization,
-            first_prenormalization=first_prenormalization,
-            last_layer_query_idx=None,
-            n_tokens=None,
-            kv_compression_ratio=kv_compression_ratio,
-            kv_compression_sharing=kv_compression_sharing,
-            head_activation=head_activation,
-            head_normalization=head_normalization,
-            d_out=hidden_features,
-            projection=False,
-            additive_attention=additive_attention,
-            share_qv_weights=share_qv_weights,
-        )
-
-        self.head = FT_Transformer.Head(
-            d_in=in_features,
-            d_out=num_classes,
-            bias=True,
-            activation=head_activation,
-            normalization=head_normalization,
-        )
-
-        self.cls_token = CLSToken(
-            d_token=in_features,
-            initialization="uniform",
-        )
-
-        self.out_features = in_features
-
-        # init weights
-        self.adapter.apply(init_weights)
-        self.head.apply(init_weights)
-
-        self.prefix = prefix
-
-        self.name_to_id = self.get_layer_ids()
-        self.head_layer_names = [n for n, layer_id in self.name_to_id.items() if layer_id == 0]
-
-    def forward(
-        self,
-        batch: dict,
-    ):
-        multimodal_features = []
-        output = {}
-        for per_model, per_adapter in zip(self.model, self.adapter):
-            per_output = run_model(per_model, batch)
-            multimodal_feature = per_adapter(per_output[per_model.prefix][FEATURES])
-            if multimodal_feature.ndim == 2:
-                multimodal_feature = torch.unsqueeze(multimodal_feature, dim=1)
-            multimodal_features.append(multimodal_feature)
-
-            if self.loss_weight is not None:
-                per_output[per_model.prefix].update({WEIGHT: self.loss_weight})
-                output.update(per_output)
-
-        multimodal_features = torch.cat(multimodal_features, dim=1)
-        multimodal_features = self.cls_token(multimodal_features)
-        features = self.fusion_transformer(multimodal_features)
-
-        logits = self.head(features)
-        fusion_output = {
-            self.prefix: {
-                LOGITS: logits,
-                FEATURES: features,
-            }
-        }
-        if self.loss_weight is not None:
-            fusion_output[self.prefix].update({WEIGHT: 1})
-            output.update(fusion_output)
-            return output
-        else:
-            return fusion_output
