+++ multimodal/src/autogluon/multimodal/models/fusion/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseMultimodalFusionModel
+from .fusion_mlp import MultimodalFusionMLP
+from .fusion_ner import MultimodalFusionNER
+from .fusion_transformer import MultimodalFusionTransformer
